Separate runner factory from the context factory.
